Keyboard interrupt handling
This provides pbench-tool-meister-start with a framework to consistently
and reliably handle errors that occur during the startup process,
including a KeyboardInterrupt exception. The structured cleanup recovery
mechanism from the server PUT operation has been generalized and is
leveraged here to ensure orderly shutdown.

The pbench-tool-meister-stop command is improved to be a bit more
friendly when `--interrupt` is used to signal that it's not a normal
termination (e.g., allowing that the "end" command may fail because
clients might be in running state).

Behavior of bash benchmark scripts that use Tool Meister are improved
with a change to the agent `base` interrupt handler, which now exits
rather than returning back into the benchmark code after shutting down
the Tool Meister.

It's worth noting the "unconventional" use of an f-string in a logger
output in the cleanup code rather than deferred Logger message
formatting: this is required because the cleanup code is now shared
between server ("{}" syntax) and agent ("%s" syntax). A more complex
workaround didn't seem justified as we're not ever going to be
disabling `exception` log records.
